ST-1130 Check answers adn PMD fixes
