+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/service/DocumentListService.java
@@ -7,10 +7,8 @@ import org.springframework.util.ObjectUtils;
 import uk.gov.hmcts.ccd.sdk.type.DynamicList;
 import uk.gov.hmcts.ccd.sdk.type.DynamicListElement;
 import uk.gov.hmcts.ccd.sdk.type.ListValue;
-import uk.gov.hmcts.sptribs.caseworker.model.Order;
 import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.ciccase.model.CicCase;
-import uk.gov.hmcts.sptribs.document.model.CICDocument;
 import uk.gov.hmcts.sptribs.document.model.CaseworkerCICDocument;
 import uk.gov.hmcts.sptribs.document.model.DocumentType;
 
@@ -19,6 +17,8 @@ import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
+
+import static uk.gov.hmcts.sptribs.caseworker.util.OrderDocumentListUtil.getOrderDocuments;
 
 @Service
 @Slf4j
@@ -146,35 +146,6 @@ public class DocumentListService {
             }
         }
         return hearingSummaryDocs;
-    }
-
-    public List<CaseworkerCICDocument> getOrderDocuments(CicCase cicCase) {
-        List<CaseworkerCICDocument> orderList = new ArrayList<>();
-        if (!CollectionUtils.isEmpty(cicCase.getOrderList())) {
-            for (ListValue<Order> orderListValue : cicCase.getOrderList()) {
-                if (null != orderListValue.getValue().getDraftOrder()
-                    && null != orderListValue.getValue().getDraftOrder().getTemplateGeneratedDocument()
-                    && !ObjectUtils.isEmpty(orderListValue.getValue().getDraftOrder().getTemplateGeneratedDocument().getFilename())) {
-                    CaseworkerCICDocument doc = CaseworkerCICDocument.builder()
-                        .documentLink(orderListValue.getValue().getDraftOrder().getTemplateGeneratedDocument())
-                        .documentCategory(DocumentType.TRIBUNAL_DIRECTION)
-                        .build();
-                    orderList.add(doc);
-                } else if (!CollectionUtils.isEmpty(orderListValue.getValue().getUploadedFile())) {
-                    for (ListValue<CICDocument> document : orderListValue.getValue().getUploadedFile()) {
-                        if (null != document.getValue().getDocumentLink()) {
-                            CaseworkerCICDocument doc = CaseworkerCICDocument.builder()
-                                .documentLink(document.getValue().getDocumentLink())
-                                .documentEmailContent(document.getValue().getDocumentEmailContent())
-                                .documentCategory(DocumentType.TRIBUNAL_DIRECTION)
-                                .build();
-                            orderList.add(doc);
-                        }
-                    }
-                }
-            }
-        }
-        return orderList;
     }
 
     public List<ListValue<CaseworkerCICDocument>> getAllDecisionDocuments(CaseData caseData) {
