+++ src/test/java/uk/gov/hmcts/sptribs/caseworker/service/DocumentListServiceTest.java
@@ -265,23 +265,6 @@ public class DocumentListServiceTest {
     }
 
     @Test
-    void shouldGenerateOrderDocTemp() {
-        //Given
-
-        Order order = Order.builder().draftOrder(DraftOrderCIC.builder()
-            .templateGeneratedDocument(Document.builder().filename("name").build()).build()).build();
-        ListValue<Order> orderListValue = new ListValue<>();
-        orderListValue.setValue(order);
-        CicCase cicCase = CicCase.builder().orderList(List.of(orderListValue)).build();
-
-        //When
-        List<CaseworkerCICDocument> result = documentListService.getOrderDocuments(cicCase);
-
-        //Then
-        assertThat(result).isNotNull();
-    }
-
-    @Test
     void shouldGenerateOrderAll() {
         //Given
         Order order = Order.builder().draftOrder(DraftOrderCIC.builder()
@@ -297,23 +280,5 @@ public class DocumentListServiceTest {
         assertThat(result).isNotNull();
     }
 
-    @Test
-    void shouldGenerateOrderDocUpload() {
-        //Given
-
-        CICDocument document = CICDocument.builder().documentLink(Document.builder().filename("name").build()).build();
-        ListValue<CICDocument> cicDocumentListValue = new ListValue<>();
-        cicDocumentListValue.setValue(document);
-        Order order = Order.builder().uploadedFile(List.of(cicDocumentListValue)).build();
-        ListValue<Order> orderListValue = new ListValue<>();
-        orderListValue.setValue(order);
-        CicCase cicCase = CicCase.builder().orderList(List.of(orderListValue)).build();
-
-        //When
-        List<CaseworkerCICDocument> result = documentListService.getOrderDocuments(cicCase);
-
-        //Then
-        assertThat(result).isNotNull();
-    }
 
 }
